Rename client api package from "processor" to "service" (for hbase and elasticsearch)
