+++ logisland-plugins/logisland-elasticsearch-plugin/src/main/java/com/hurence/logisland/processor/elasticsearch/MultiGetElasticsearch.java
@@ -6,9 +6,9 @@ import com.hurence.logisland.annotation.documentation.Tags;
 import com.hurence.logisland.component.PropertyDescriptor;
 import com.hurence.logisland.processor.ProcessContext;
 import com.hurence.logisland.processor.ProcessError;
-import com.hurence.logisland.processor.elasticsearch.multiGet.InvalidMultiGetQueryRecordException;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetQueryRecord;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetResponseRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.InvalidMultiGetQueryRecordException;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetQueryRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetResponseRecord;
 import com.hurence.logisland.record.*;
 import com.hurence.logisland.validator.StandardValidators;
