+++ logisland-plugins/logisland-hbase-plugin/src/main/java/com/hurence/logisland/processor/hbase/FetchHBaseRow.java
@@ -12,9 +12,10 @@ import com.hurence.logisland.processor.AbstractProcessor;
 import com.hurence.logisland.processor.ProcessContext;
 import com.hurence.logisland.processor.ProcessError;
 import com.hurence.logisland.processor.ProcessException;
-import com.hurence.logisland.processor.hbase.scan.Column;
-import com.hurence.logisland.processor.hbase.scan.ResultCell;
-import com.hurence.logisland.processor.hbase.scan.ResultHandler;
+import com.hurence.logisland.service.hbase.HBaseClientService;
+import com.hurence.logisland.service.hbase.scan.Column;
+import com.hurence.logisland.service.hbase.scan.ResultCell;
+import com.hurence.logisland.service.hbase.scan.ResultHandler;
 import com.hurence.logisland.record.Record;
 import com.hurence.logisland.serializer.*;
 import com.hurence.logisland.validator.StandardValidators;
