+++ logisland-plugins/logisland-hbase-plugin/src/main/java/com/hurence/logisland/processor/hbase/util/RowSerializerUtil.java
@@ -2,7 +2,7 @@
 package com.hurence.logisland.processor.hbase.util;
 
 
-import com.hurence.logisland.processor.hbase.scan.ResultCell;
+import com.hurence.logisland.service.hbase.scan.ResultCell;
 
 import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
