+++ logisland-plugins/logisland-hbase-plugin/src/test/java/com/hurence/logisland/processor/hbase/HBaseTestUtil.java
@@ -1,7 +1,7 @@
 package com.hurence.logisland.processor.hbase;
 
-import com.hurence.logisland.processor.hbase.put.PutColumn;
-import com.hurence.logisland.processor.hbase.put.PutRecord;
+import com.hurence.logisland.service.hbase.put.PutColumn;
+import com.hurence.logisland.service.hbase.put.PutRecord;
 
 import static org.junit.Assert.assertTrue;
