+++ logisland-plugins/logisland-hbase-plugin/src/test/java/com/hurence/logisland/processor/hbase/TestPutHBaseCell.java
@@ -2,8 +2,8 @@
 package com.hurence.logisland.processor.hbase;
 
 import com.hurence.logisland.component.InitializationException;
-import com.hurence.logisland.processor.hbase.put.PutColumn;
-import com.hurence.logisland.processor.hbase.put.PutRecord;
+import com.hurence.logisland.service.hbase.put.PutColumn;
+import com.hurence.logisland.service.hbase.put.PutRecord;
 import com.hurence.logisland.record.Record;
 import com.hurence.logisland.record.RecordUtils;
 import com.hurence.logisland.serializer.KryoSerializer;
