+++ logisland-plugins/logisland-hbase-plugin/src/test/java/com/hurence/logisland/processor/hbase/io/TestJsonRowSerializer.java
@@ -1,7 +1,7 @@
 package com.hurence.logisland.processor.hbase.io;
 
 
-import com.hurence.logisland.processor.hbase.scan.ResultCell;
+import com.hurence.logisland.service.hbase.scan.ResultCell;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
