+++ logisland-services/logisland-elasticsearch-client-service-api/src/main/java/com/hurence/logisland/service/elasticsearch/ElasticsearchClientService.java
@@ -1,4 +1,4 @@
-package com.hurence.logisland.processor.elasticsearch;
+package com.hurence.logisland.service.elasticsearch;
 
 
 import com.hurence.logisland.annotation.documentation.CapabilityDescription;
@@ -6,8 +6,8 @@ import com.hurence.logisland.annotation.documentation.Tags;
 import com.hurence.logisland.component.AllowableValue;
 import com.hurence.logisland.component.PropertyDescriptor;
 import com.hurence.logisland.controller.ControllerService;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetQueryRecord;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetResponseRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetQueryRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetResponseRecord;
 import com.hurence.logisland.record.Record;
 import com.hurence.logisland.validator.StandardValidators;
 import com.hurence.logisland.validator.ValidationResult;
