+++ logisland-services/logisland-elasticsearch-client-service-api/src/main/java/com/hurence/logisland/service/elasticsearch/multiGet/MultiGetQueryRecord.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hurence.logisland.processor.elasticsearch.multiGet;
+package com.hurence.logisland.service.elasticsearch.multiGet;
 
 
 import java.util.List;
