+++ logisland-services/logisland-elasticsearch_2_3_3-client-service/src/main/java/com/hurence/logisland/service/elasticsearch/Elasticsearch_2_3_3_ClientService.java
@@ -9,9 +9,8 @@ import com.hurence.logisland.component.PropertyDescriptor;
 import com.hurence.logisland.controller.AbstractControllerService;
 import com.hurence.logisland.controller.ControllerServiceInitializationContext;
 import com.hurence.logisland.processor.ProcessException;
-import com.hurence.logisland.processor.elasticsearch.ElasticsearchClientService;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetQueryRecord;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetResponseRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetQueryRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetResponseRecord;
 import com.hurence.logisland.record.Record;
 import org.apache.commons.lang3.StringUtils;
 import org.elasticsearch.action.ActionRequestValidationException;
