+++ logisland-services/logisland-elasticsearch_2_4_0-client-service/src/test/java/com/hurence/logisland/service/elasticsearch/TestElasticsearch_2_4_0_ClientService.java
@@ -20,10 +20,9 @@ import com.hurence.logisland.component.InitializationException;
 import com.hurence.logisland.component.PropertyDescriptor;
 import com.hurence.logisland.controller.ControllerServiceInitializationContext;
 import com.hurence.logisland.processor.ProcessException;
-import com.hurence.logisland.processor.elasticsearch.ElasticsearchClientService;
-import com.hurence.logisland.processor.elasticsearch.multiGet.InvalidMultiGetQueryRecordException;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetQueryRecord;
-import com.hurence.logisland.processor.elasticsearch.multiGet.MultiGetResponseRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.InvalidMultiGetQueryRecordException;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetQueryRecord;
+import com.hurence.logisland.service.elasticsearch.multiGet.MultiGetResponseRecord;
 import com.hurence.logisland.util.runner.TestRunner;
 import com.hurence.logisland.util.runner.TestRunners;
 import org.elasticsearch.action.bulk.BulkItemResponse;
