+++ logisland-services/logisland-elasticsearch_2_4_0-client-service/src/test/java/com/hurence/logisland/service/elasticsearch/TestProcessor.java
@@ -4,7 +4,6 @@ package com.hurence.logisland.service.elasticsearch;
 import com.hurence.logisland.component.PropertyDescriptor;
 import com.hurence.logisland.processor.AbstractProcessor;
 import com.hurence.logisland.processor.ProcessContext;
-import com.hurence.logisland.processor.elasticsearch.ElasticsearchClientService;
 import com.hurence.logisland.record.Record;
 
 import java.util.ArrayList;
