+++ logisland-services/logisland-hbase-client-service-api/src/main/java/com/hurence/logisland/service/hbase/put/PutRecord.java
@@ -1,4 +1,4 @@
-package com.hurence.logisland.processor.hbase.put;
+package com.hurence.logisland.service.hbase.put;
 
 
 import com.hurence.logisland.record.Record;
