+++ logisland-services/logisland-hbase-client-service-api/src/main/java/com/hurence/logisland/service/hbase/scan/Column.java
@@ -1,4 +1,4 @@
-package com.hurence.logisland.processor.hbase.scan;
+package com.hurence.logisland.service.hbase.scan;
 
 import java.util.Arrays;
